[module5-task2]: DataBase: Move constants to common file
