+++ .gitignore
@@ -13,4 +13,4 @@ node_modules/
 
 # Папка с собранными файлами проекта
 build/
-mocks/*.json
+mocks/
