--- src/service/cli/constants/table-names.constant.ts
@@ -1,7 +1,0 @@
-export const TableNames = {
-  USERS: `users`,
-  PERMISSIONS: `permissions`,
-  ARTICLES: `articles`,
-  CATEGORIES: `categories`,
-  ARTICLES_CATEGORIES: `articles_categories`,
-} as const;
